# bin/docs/swagger.yaml
swagger: '2.0'
info:
  version: "1.0.0"
  title: Doppler App REST API
host: localhost:8080
basePath: /api/v1
schemes:
  - http
  - https
produces:
  - application/json
consumes:
  - application/json
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
paths:
...
  /events/{eventId}:
    get:
      summary: get an event
      tags:
        - Event
      security:
        - Bearer: []
      parameters:
        - in: path
          name: eventId
          required: true
          type: string
          description: the event UUID
      responses:
        200:
          description: successful operation
          schema:
            $ref: "#/definitions/Event"
        400:
          description: invalid request
          schema:
            $ref: "#/definitions/ErrorMessage"
        401:
          description: action is not allowed for anonymous | invalid or expired access token
          schema:
            $ref: "#/definitions/ErrorMessage"
        404:
          description: the event does not exist
          schema:
            $ref: "#/definitions/ErrorMessage"
        500:
          description: server error
          schema:
            $ref: "#/definitions/ErrorMessage"
...
  /events/{eventId}/register:
    post:
      summary: register the current logged-in user to an event
      description: >
        - Make sure that the user has not registered to the event yet
        
        - Add a new `UserEvent`
        
        - Add a new `UserRewardPoint` for registering to the event, the points should be taken from the `Event.pointsForRegistering`
        
        - Add the points to the `User.points`
        
        - Return the `UserRewardPoint`
      tags:
        - User Event
      security:
        - Bearer: []
      parameters:
        - in: path
          name: eventId
          required: true
          type: string
          description: the event UUID
      responses:
        200:
          description: successful operation
          schema:
            $ref: "#/definitions/UserRewardPoint"
        400:
          description: invalid request
          schema:
            $ref: "#/definitions/ErrorMessage"
        401:
          description: action is not allowed for anonymous | invalid or expired access token
          schema:
            $ref: "#/definitions/ErrorMessage"
        404:
          description: the event does not exist
          schema:
            $ref: "#/definitions/ErrorMessage"
        500:
          description: server error
          schema:
            $ref: "#/definitions/ErrorMessage"
  /events/{eventId}/attendees:
    get:
      summary: get all attendees of a specific event
      tags:
        - User Event
      security:
        - Bearer: []
      parameters:
        - in: path
          name: eventId
          required: true
          type: string
          description: the event UUID
      responses:
        200:
          description: successful operation
          schema:
            type: array
            items:
              $ref: "#/definitions/User"
        400:
          description: invalid request
          schema:
            $ref: "#/definitions/ErrorMessage"
        401:
          description: action is not allowed for anonymous | invalid or expired access token
          schema:
            $ref: "#/definitions/ErrorMessage"
        404:
          description: the event does not exist
          schema:
            $ref: "#/definitions/ErrorMessage"
        500:
          description: server error
          schema:
            $ref: "#/definitions/ErrorMessage"
  /events/{eventId}/scanTicket:
    post:
      summary: confirm the atttendance to the event
      description: >
        - NOTE this endpoint will use a special configurable bearer token to authenticate (`token.m2m`)
      
        - Make sure that the user has not scan the ticket yet (UserEvent.ticketScanned = false)
        
        - Update the `UserEvent.ticketScanned` to true
        
        - Add a new `UserRewardPoint` for attending the event, the points should be taken from the `Event.pointsForScanningTicket`
        
        - Add the points to the `User.points`
        
        - Return the `UserRewardPoint`
      tags:
        - User Event     
      security:
        - Bearer: []
      parameters:
        - in: path
          name: eventId
          required: true
          type: string
          description: the event UUID
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              userId:
                type: string
      responses:
        200:
          description: successful operation
          schema:
            $ref: "#/definitions/UserRewardPoint"
        400:
          description: invalid request
          schema:
            $ref: "#/definitions/ErrorMessage"
        404:
          description: the event does not exist
          schema:
            $ref: "#/definitions/ErrorMessage"
        500:
          description: server error
          schema:
            $ref: "#/definitions/ErrorMessage"
...
  /events/{eventId}/survey:
    get:
      summary: get the survey questions of the event for the logged-in user
      tags:
        - Survey
      security:
        - Bearer: []
      parameters:
        - in: path
          name: eventId
          type: string
          required: true
      responses:
        200:
          description: successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/SurveyQuestion'
        400:
          description: invalid request
          schema:
            $ref: "#/definitions/ErrorMessage"
        401:
          description: action is not allowed for anonymous | invalid or expired access token
          schema:
            $ref: "#/definitions/ErrorMessage"
        403:
          description: the user is not allowed to get survey questions of the event he/she has not attended
          schema:
            $ref: "#/definitions/ErrorMessage"
        404:
          description: the event does not exist
          schema:
            $ref: "#/definitions/ErrorMessage"
        500:
          description: server error
          schema:
            $ref: "#/definitions/ErrorMessage"
            
    post:
      summary: submit the survey answers of the event for the logged-in user
      description: >
        - Add a new `UserRewardPoint` for completing the survey, the points should be got from the event or configuration
      
        - Add the points to the `User.points`
        
        - Return the `UserRewardPoint`
      tags:
        - Survey
      security:
        - Bearer: []
      parameters:
        - in: path
          name: eventId
          type: string
          required: true
        - in: body
          name: body
          required: true
          schema:
            type: array
            items:
              $ref: '#/definitions/SurveyAnswer'
      responses:
        200:
          description: successful operation
          schema:
            $ref: '#/definitions/UserRewardPoint'
        400:
          description: invalid request
          schema:
            $ref: "#/definitions/ErrorMessage"
        401:
          description: action is not allowed for anonymous | invalid or expired access token
          schema:
            $ref: "#/definitions/ErrorMessage"
        403:
          description: the user is not allowed to submit survey answers for the event he/she has not attended
          schema:
            $ref: "#/definitions/ErrorMessage"
        404:
          description: the event does not exist
          schema:
            $ref: "#/definitions/ErrorMessage"
        500:
          description: server error
          schema:
            $ref: "#/definitions/ErrorMessage"
...
definitions:
  ErrorMessage:
    type: object
    properties:
      message:
        type: string
...
  User:
    type: object
    properties:
      id:
        type: string
        description: the UUID
      email:
        type: string
      fullName:
        type: string
        description: synced from SAML
      division:
        type: string
        description: synced from SAML
      position:
        type: string
        description: synced from SAML
      location:
        type: string
        description: synced from SAML
      phoneNumber:
        type: string
        description: synced from SAML
      photoUrl:
        type: string
        description: synced from SAML
      isNew:
        type: boolean
        description: the flag to indicate that the user is a new or an existing user, if the user is a new user, client side should display the sign-up screens so that the user can select the topic references
      points:
        type: integer
        minimum: 0
        description: the current points
...
  UserRewardPoint:
    type: object
    description: the history of points earned by the user
    properties:
      id:
        type: string
        description: the UUID, should not be present in the request body
      description:
        type: string
      points:
        type: number
        format: integer
        minimum: 1
      createdAt:
        type: string
        format: date
        description: the created date time, generated by the system and should not be present in the request body
      userId:
        type: string
        description: the user UUID
    required:
      - description
      - points
      - createdAt
      - userId
  Event:
    type: object
    properties:
      id:
        type: string
        description: the UUID, should not be present in the request body
      title:
        type: string
      description:
        type: string
      pointsForRegistering:
        type: number
        format: integer
        minimum: 1
        description: the points that each attendee earns for registering to the event
      pointsForScanningTicket:
        type: number
        format: integer
        minimum: 1
        description: the points that each attendee earns for scanning ticket
      topicId:
        type: string
      location:
        type: string
      currentBooking:
        type: integer
        description: the current booking
      maxCapacity:
        type: integer
        description: the max capacity
      waitingCapacity:
        type: integer
        description: the waiting capacity
      tags:
        type: array
        items:
          type: string
      sessions:
        type: array
        items:
          $ref: '#/definitions/EventSession'
      portraitImageUrl:
        type: string
        description: the portrait image url (to be displayed in the event list)
      landscapeImageUrl:
        type: string
        description: the landscape image url (to be displayed in the event detail page)
      isMyEvent:
        type: boolean
        description: the flag to indicate whether the current logged-in user has registered to the event
  EventSession:
    type: object
    properties:
      id:
        type: string
        description: the UUID, should not be present in the request body
      title:
        type: string
      location:
        type: string
      start:
        type: string
        format: date
      end:
        type: string
        format: date
      speaker:
        type: object
        properties:
          name:
            type: string
          photoUrl:
            type: string
...
  SurveyQuestion:
    type: object
    properties:
      id:
        type: string
      question:
        type: string
      type:
        type: string
        enum:
          - text
          - singleChoice
          - multipleChoice
      surveyAnswerOptions:
        type: array
        items:
          $ref: '#/definitions/SurveyAnswerOption'
        description: 0 items if the type is `text`
    required:
      - question
      - type
  SurveyAnswerOption:
    type: object
    properties:
      id:
        type: string
      answerOption:
        type: string
    required:
      - surveyQuestionId
      - answerOption
  SurveyAnswer:
    type: object
    properties:
      id:
        type: string
        description: the survey answer UUID, should not be present in the request body
      surveyQuestionId:
        type: string
      selectedSurveyAnswerOptionId:
        type: string
        description: null if the question type is `text`
      textAnswer:
        type: string
        description: not null if the question type is `text`
    required:
      - surveyQuestionId
...